Add DateHelper
Add OnlineRepository with function for retrieving historical prices from NetworkLayer

+++ CoinTracker/Models/Price.swift
@@ -8,6 +8,6 @@ import Foundation
 
 struct Price {
     let currency: String
-    let value: Decimal
+    let value: Double
     let date: String
 }
